+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Add support to retrieve SSL Certificate from IAM to support 4096-bit RSA keys or EC keys
+
 ### Changed
 - Upgrade ruby_aem to 3.8.0
+- Lockdown puppet-agent version to 5.5.20 [#218]
+- Upgrade aem_resources to 7.1.0
 
 ### Added
 - Added Granite SSL method [shinesolutions/puppet-aem-curator#220]